requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt,argon2]>=1.7.4
argon2-cffi>=23.1.0
python-multipart>=0.0.6

# File Processing
pandas>=2.1.4
openpyxl>=3.1.2
xlrd>=2.0.1
python-magic>=0.4.27

# Template Engine
jinja2>=3.1.3

# PDF Generation
playwright>=1.41.0

# Storage
boto3>=1.34.0

# Background Tasks
celery[redis]>=5.3.6
redis>=5.0.1

# Utilities
httpx>=0.26.0
python-dateutil>=2.8.2
email-validator>=2.1.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
